Update ExecuteButton and Combobox components
Add call to update all combobox that contain all files in the project. Now every time the user download a new file or create a new one by filecreation package, both components comboBox and FileListCombox are loaded and updated with the new files.

ExecutButton on FileCreation Package was update to make it not checked every time a function is executed and finished.