Update HycomPackage parallel download process
Update download process to make it possible to verify the progress using a new variable called processing_ that can be boolean (True if the process is successful or False is the process is stopped).

That way, it will be possible to the main code verify what function should be called after the download processing and how to proceed.