Update CopernicusPackage parallel download process
Update download process to make it possible to verify the progress using a new variable called processing_ that can be boolean (True if the process is successful, False is the process is stopped or str if the process is interrupted by any error).

That verification is possible using multiprocessing library.

That way, it will be possible to the main code verify what function should be called after the download processing and how to proceed.