Update ViewPage Variables Identification
Update how to identify variables from file selected to filter which variable user wish to see.

If there is more than one variable with more than one dimension on file selected, a frame with radio buttons will be created so that the user can choose which variable they wish to visualize.

If there is only one variable on file with more than one dimension, the frame created will only show a message indicating that the file has only one variable.